+++ chess4j-native/src/main/native/com/jamesswafford/chess4j/hash/PawnTranspositionTable_resizeNative.c
@@ -11,10 +11,10 @@ extern hash_table_t phtbl;
 /*
  * Class:     com_jamesswafford_chess4j_hash_PawnTranspositionTable
  * Method:    resizeNative
- * Signature: (I)V
+ * Signature: (J)V
  */
 JNIEXPORT void JNICALL Java_com_jamesswafford_chess4j_hash_PawnTranspositionTable_resizeNative
-  (JNIEnv *env, jobject UNUSED(phtable), jint size_bytes)
+  (JNIEnv *env, jobject UNUSED(phtable), jlong size_bytes)
 {
     /* ensure the static library is initialized */
     if (!p4_initialized) 
@@ -24,7 +24,7 @@ JNIEXPORT void JNICALL Java_com_jamesswafford_chess4j_hash_PawnTranspositionTabl
     }
     
 
-    int retval = resize_hash_table(&phtbl, (uint32_t) size_bytes);
+    int retval = resize_hash_table(&phtbl, (uint64_t) size_bytes);
     if (0 != retval)
     {
         (*env)->ThrowNew(env, IllegalStateException, "Failed to (re)allocate hash table");
