+++ chess4j-native/src/main/native/com/jamesswafford/chess4j/hash/TranspositionTable_resizeNative.c
@@ -11,10 +11,10 @@ extern hash_table_t htbl;
 /*
  * Class:     com_jamesswafford_chess4j_hash_TranspositionTable
  * Method:    resizeNative
- * Signature: (I)V
+ * Signature: (J)V
  */
 JNIEXPORT void JNICALL Java_com_jamesswafford_chess4j_hash_TranspositionTable_resizeNative
-  (JNIEnv *env, jobject UNUSED(htable), jint size_bytes)
+  (JNIEnv *env, jobject UNUSED(htable), jlong size_bytes)
 {
     /* ensure the static library is initialized */
     if (!p4_initialized) 
@@ -24,7 +24,7 @@ JNIEXPORT void JNICALL Java_com_jamesswafford_chess4j_hash_TranspositionTable_re
     }
     
 
-    int retval = resize_hash_table(&htbl, (uint32_t) size_bytes);
+    int retval = resize_hash_table(&htbl, (uint64_t) size_bytes);
     if (0 != retval)
     {
         (*env)->ThrowNew(env, IllegalStateException, "Failed to (re)allocate hash table");
